agent custom counters & refactor
