+++ agent/provider/src/events.rs
@@ -1,4 +1,4 @@
-use crate::market::Presets;
+use crate::config::presets::Presets;
 
 #[derive(Clone, Debug)]
 pub enum Event {
