+++ agent/provider/src/execution.rs
@@ -1,10 +1,12 @@
-mod exeunit_instance;
-mod exeunits_registry;
-mod task;
-mod task_runner;
-
-pub use self::exeunits_registry::{ExeUnitDesc, ExeUnitsRegistry};
 pub use task_runner::{
     ActivityDestroyed, CreateActivity, DestroyActivity, GetExeUnit, GetOfferTemplates, Shutdown,
     TaskRunner, TaskRunnerConfig, TerminateActivity, UpdateActivity,
 };
+
+pub use self::registry::Configuration;
+pub use self::registry::{ExeUnitDesc, ExeUnitsRegistry};
+
+mod exeunit_instance;
+mod registry;
+mod task;
+mod task_runner;
