+++ agent/provider/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod cli;
+pub mod config;
 pub mod dir;
 pub mod display;
 pub mod events;
@@ -6,11 +7,10 @@ pub mod execution;
 pub mod hardware;
 pub mod market;
 pub mod payments;
-pub mod preset_cli;
 pub mod provider_agent;
 pub mod signal;
 pub mod startup_config;
 pub mod tasks;
 
-pub use provider_agent::GlobalsState;
+pub use config::globals::GlobalsState;
 pub use startup_config::ReceiverAccount;
