+++ agent/provider/src/market/presets.rs
@@ -1,62 +1,17 @@
-use crate::events::Event;
-use crate::startup_config::FileMonitor;
-use anyhow::{anyhow, Error, Result};
-use serde::{Deserialize, Serialize};
-use std::collections::{HashMap, HashSet};
-use std::convert::TryFrom;
+use std::collections::HashMap;
 use std::fmt;
+use std::fmt::Formatter;
 use std::path::Path;
 use std::sync::{Arc, Mutex};
-use strum::IntoEnumIterator;
-use strum_macros::EnumIter;
+
+use anyhow::{anyhow, Result};
+use serde::{Deserialize, Serialize};
 use tokio::sync::watch;
-use ya_utils_path::SwapSave;
 
-#[derive(Serialize, Deserialize, Clone, Debug, Hash, Eq, PartialEq, EnumIter)]
-#[serde(rename_all = "snake_case")]
-#[non_exhaustive]
-pub enum Coefficient {
-    Duration,
-    Cpu,
-    Initial,
-}
-
-impl Coefficient {
-    #[inline]
-    pub fn variants() -> impl Iterator<Item = Coefficient> {
-        Self::iter()
-    }
-
-    pub fn to_property(&self) -> Option<&str> {
-        let property = match self {
-            Coefficient::Duration => "golem.usage.duration_sec",
-            Coefficient::Cpu => "golem.usage.cpu_sec",
-            Coefficient::Initial => return None,
-        };
-        Some(property)
-    }
-
-    pub fn to_readable(&self) -> &str {
-        match self {
-            Coefficient::Duration => "Duration",
-            Coefficient::Cpu => "CPU",
-            Coefficient::Initial => "Init price",
-        }
-    }
-}
-
-impl<'s> TryFrom<&'s str> for Coefficient {
-    type Error = Error;
-
-    fn try_from(value: &'s str) -> Result<Self> {
-        match value {
-            "Init price" => Ok(Coefficient::Initial),
-            "Duration" => Ok(Coefficient::Duration),
-            "CPU" => Ok(Coefficient::Cpu),
-            _ => Err(anyhow!("Invalid coefficient: {}", value)),
-        }
-    }
-}
+pub use crate::config::presets::Presets;
+use crate::events::Event;
+use crate::execution::ExeUnitsRegistry;
+use crate::startup_config::FileMonitor;
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
 #[serde(rename_all = "kebab-case")]
@@ -65,12 +20,20 @@ pub struct Preset {
     pub name: String,
     pub exeunit_name: String,
     pub pricing_model: String,
-    pub usage_coeffs: HashMap<Coefficient, f64>,
+    pub initial_price: f64,
+    pub usage_coeffs: HashMap<String, f64>,
 }
 
 impl Preset {
     pub fn get_initial_price(&self) -> Option<f64> {
-        self.usage_coeffs.get(&Coefficient::Initial).cloned()
+        Some(self.initial_price)
+    }
+
+    pub fn display<'a, 'b>(&'a self, registry: &'b ExeUnitsRegistry) -> PresetDisplay<'a, 'b> {
+        PresetDisplay {
+            preset: self,
+            registry,
+        }
     }
 }
 
@@ -80,113 +43,6 @@ pub struct PresetManager {
     monitor: Option<FileMonitor>,
     sender: Option<watch::Sender<Event>>,
     receiver: watch::Receiver<Event>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "kebab-case")]
-pub struct Presets {
-    pub active: Vec<String>,
-    pub presets: HashMap<String, Preset>,
-}
-
-impl Presets {
-    pub fn load_from_file<P: AsRef<Path>>(presets_file: P) -> Result<Presets> {
-        let path = presets_file.as_ref();
-        log::debug!("Loading presets from: {}", path.display());
-        let json = std::fs::read_to_string(path)?;
-        let presets: Presets = serde_json::from_str::<PresetsFile>(json.as_str())
-            .map_err(|e| anyhow!("Can't deserialize Presets from file {:?}: {}", path, e))?
-            .into();
-
-        match presets.active.is_empty() {
-            false => presets.active.iter().try_for_each(|name| {
-                presets
-                    .presets
-                    .get(name)
-                    .ok_or(anyhow!("Invalid active preset: {:?}", name))
-                    .map(|_| ())
-            })?,
-            _ => return Err(anyhow!("No active presets")),
-        }
-
-        Ok(presets)
-    }
-
-    pub fn save_to_file(&self, presets_file: &Path) -> Result<()> {
-        let json = serde_json::to_string_pretty(&PresetsFile::from(self))
-            .map_err(|error| anyhow!("Failed to serialize Presets: {}", error))?;
-        presets_file.swap_save(json).map_err(|error| {
-            anyhow!(
-                "Failed to save Presets to file {}, error: {}.",
-                presets_file.display(),
-                error
-            )
-        })?;
-        Ok(())
-    }
-
-    pub fn diff(&self, other: &Presets) -> (Vec<String>, Vec<String>) {
-        let mut updated = HashSet::new();
-        let mut removed = HashSet::new();
-
-        self.active.iter().for_each(|n| {
-            if !other.active.contains(n) {
-                removed.insert(n.clone());
-            }
-        });
-        self.presets
-            .iter()
-            .for_each(|(n, p)| match other.presets.get(n) {
-                Some(preset) => {
-                    if preset != p {
-                        updated.insert(n.clone());
-                    }
-                }
-                _ => {
-                    removed.insert(n.clone());
-                }
-            });
-
-        (updated.into_iter().collect(), removed.into_iter().collect())
-    }
-}
-
-impl Default for Presets {
-    fn default() -> Self {
-        Presets {
-            active: Vec::new(),
-            presets: HashMap::new(),
-        }
-    }
-}
-
-// FIXME: drop Preset::name so PresetsState can be serialized without conversion
-#[derive(Serialize, Deserialize, Debug)]
-struct PresetsFile {
-    active: Vec<String>,
-    presets: Vec<Preset>,
-}
-
-impl From<PresetsFile> for Presets {
-    fn from(presets_file: PresetsFile) -> Self {
-        Presets {
-            active: presets_file.active,
-            presets: presets_file
-                .presets
-                .into_iter()
-                .map(|p| (p.name.clone(), p))
-                .collect(),
-        }
-    }
-}
-
-impl<'p> From<&'p Presets> for PresetsFile {
-    fn from(presets: &'p Presets) -> Self {
-        PresetsFile {
-            active: presets.active.clone(),
-            presets: presets.presets.values().cloned().collect(),
-        }
-    }
 }
 
 impl PresetManager {
@@ -338,9 +194,6 @@ impl PresetManager {
     pub fn deactivate(&mut self, name: &String) -> Result<()> {
         let mut state = self.state.lock().unwrap();
         if let Some(idx) = state.active.iter().position(|n| name == n) {
-            if state.active.len() == 1 {
-                return Err(anyhow!("Cannot remove the last active preset: {:?}", name));
-            }
             state.active.remove(idx);
             return Ok(());
         }
@@ -364,16 +217,11 @@ impl Default for PresetManager {
 impl Default for Preset {
     // FIXME: sane defaults
     fn default() -> Self {
-        let usage_coeffs = vec![
-            (Coefficient::Duration, 0.1),
-            (Coefficient::Cpu, 1.0),
-            (Coefficient::Initial, 1.0),
-        ]
-        .into_iter()
-        .collect();
+        let usage_coeffs = Default::default();
 
         Preset {
             name: "default".to_string(),
+            initial_price: 0.0,
             exeunit_name: "wasmtime".to_string(),
             pricing_model: "linear".to_string(),
             usage_coeffs,
@@ -390,38 +238,57 @@ impl PartialEq for Preset {
     }
 }
 
-impl fmt::Display for Preset {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let align = 20;
-        let align_coeff = align - 4; // Minus indent.
+pub struct PresetDisplay<'a, 'b> {
+    preset: &'a Preset,
+    registry: &'b ExeUnitsRegistry,
+}
 
-        write!(f, "{:width$}{}\n", "Name:", self.name, width = align)?;
-        write!(
-            f,
-            "{:width$}{}\n",
-            "ExeUnit:",
-            self.exeunit_name,
-            width = align
-        )?;
-        write!(
-            f,
-            "{:width$}{}\n",
-            "Pricing model:",
-            self.pricing_model,
-            width = align
-        )?;
-        write!(f, "{}\n", "Coefficients:")?;
-
-        for (name, coeff) in self.usage_coeffs.iter() {
-            write!(
-                f,
-                "    {:width$}{} GLM\n",
-                name.to_readable(),
-                coeff,
-                width = align_coeff
-            )?;
-        }
-
-        Ok(())
+impl<'a, 'b> fmt::Display for PresetDisplay<'a, 'b> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        display_preset(f, self.preset, self.registry)
     }
+}
+
+fn display_preset(
+    f: &mut fmt::Formatter,
+    preset: &Preset,
+    registry: &ExeUnitsRegistry,
+) -> fmt::Result {
+    let align = 20;
+    let align_coeff = align - 4; // Minus indent.
+
+    write!(f, "{:width$}{}\n", "Name:", preset.name, width = align)?;
+    write!(
+        f,
+        "{:width$}{}\n",
+        "ExeUnit:",
+        preset.exeunit_name,
+        width = align
+    )?;
+    write!(
+        f,
+        "{:width$}{}\n",
+        "Pricing model:",
+        preset.pricing_model,
+        width = align
+    )?;
+    write!(f, "{}\n", "Coefficients:")?;
+
+    let exe_unit = registry.find_exeunit(&preset.exeunit_name).ok();
+
+    for (name, coeff) in preset.usage_coeffs.iter() {
+        let price_desc = exe_unit
+            .as_ref()
+            .and_then(|e| e.coefficient_name(&name))
+            .unwrap_or_else(|| name.to_string());
+        write!(
+            f,
+            "    {:width$}{} GLM\n",
+            price_desc,
+            coeff,
+            width = align_coeff
+        )?;
+    }
+
+    Ok(())
 }
