+++ agent/provider/src/payments/pricing.rs
@@ -3,11 +3,11 @@ use bigdecimal::{BigDecimal, FromPrimitive};
 use serde_json::json;
 
 use ya_agreement_utils::ComInfo;
-use ya_client_model::{payment::Account, NodeId};
+use ya_client::model::{payment::Account, NodeId};
 use ya_core_model::payment::local::NetworkName;
 
 use super::model::{PaymentDescription, PaymentModel};
-use crate::market::presets::{Coefficient, Preset};
+use crate::market::presets::Preset;
 
 #[derive(Clone, Debug)]
 pub struct AccountView {
@@ -27,10 +27,10 @@ impl From<Account> for AccountView {
 }
 
 pub trait PricingOffer {
-    fn prices(&self, preset: &Preset) -> Vec<(Coefficient, f64)>;
+    fn prices(&self, preset: &Preset) -> Vec<(String, f64)>;
     fn build(
         &self,
-        accounts: &Vec<AccountView>,
+        accounts: &[AccountView],
         initial_price: f64,
         prices: Vec<(String, f64)>,
     ) -> Result<ComInfo>;
@@ -53,7 +53,8 @@ impl PaymentModel for LinearPricing {
                 .map(|(coeff, usage_value)| coeff * usage_value)
                 .sum::<f64>();
 
-        BigDecimal::from_f64(cost).ok_or(anyhow!("Failed to convert to BigDecimal: {}", cost))
+        BigDecimal::from_f64(cost)
+            .ok_or_else(|| anyhow!("Failed to convert to BigDecimal: {}", cost))
     }
 
     fn expected_usage_len(&self) -> usize {
@@ -90,25 +91,18 @@ impl LinearPricingOffer {
     #[allow(unused)]
     pub fn interval(mut self, seconds: f64) -> Self {
         self.interval = seconds;
-        return self;
+        self
     }
 }
 
 impl PricingOffer for LinearPricingOffer {
-    fn prices(&self, preset: &Preset) -> Vec<(Coefficient, f64)> {
-        Coefficient::variants()
-            .into_iter()
-            .filter(|c| c != &Coefficient::Initial)
-            .filter_map(|c| match preset.usage_coeffs.get(&c) {
-                Some(v) => Some((c, *v)),
-                None => None,
-            })
-            .collect()
+    fn prices(&self, preset: &Preset) -> Vec<(String, f64)> {
+        preset.usage_coeffs.clone().into_iter().collect()
     }
 
     fn build(
         &self,
-        accounts: &Vec<AccountView>,
+        accounts: &[AccountView],
         initial_price: f64,
         prices: Vec<(String, f64)>,
     ) -> Result<ComInfo> {
